+++ i2cmeter2/fftfunctions.h
@@ -3,7 +3,7 @@
 #ifndef fftfunctions_h
 #define fftfunctions_h
 
-double CalculateCoeff(int freqIndex, int sampleSize);
+double CalculateCoeff(uint8_t freqIndex, int sampleSize);
 double GetMagnatude(double coefficient, int sampleSize, int *samples);
 int GrepADC(int readSampleCount, int *readArray);
